Split src/preview/proxy.ts into proxy/ subdirectory (warren-b902)
Splits the 912-line src/preview/proxy.ts into modules under
src/preview/proxy/:

- route-match.ts: host/path/referer parsers + WARREN_API_PATH_PREFIXES
- rewrite.ts: HTML <base> injection, root-relative href/src/srcset
  rewriting, Location: rewriting, isHtmlContentType
- forward.ts: forwardToUpstream + maybeFlushLastHit + debounce constant
- responses.ts: previewError + previewUnauthorized envelopes
- types.ts: PreviewProxyConfig + Deps shapes
- index.ts: createPreviewProxyHandler orchestrator + public re-exports

The 1593-line proxy.test.ts is split alongside into route-match,
rewrite, handler-subdomain, handler-path, handler-path-rewrites, and
handler-path-referer test files. A new test-helpers.ts exposes
setupProxyEnv() so each suite's beforeEach is one line instead of
repeated db/repos/auth/run/attachPreview boilerplate.

External importers (src/server/main.ts,
src/server/handlers/runs.preview.test.ts) now resolve
../preview/proxy/index.ts.

Removes proxy.ts + proxy.test.ts from scripts/file-size-budgets.json
(every new file is well under the 500-line cap). Replaces proxy.ts
with proxy/index.ts + proxy/rewrite.ts in biome.json's
noExcessiveCognitiveComplexity override — matchAttributeAt and
indexOfAsciiCaseInsensitive remain byte-level scanners that trip the
ceiling but their public API surface is unchanged.

bun run check:all green (97 proxy tests pass, lint + typecheck +
duplicates + bundle-size + ci-parity all clean).
