+++ .mulch/expertise/preview.jsonl
@@ -1,3 +1,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-26T06:09:51.861Z","evidence":{"seeds":"warren-49d9","commit":"039f3f9fa8327d20181a6141be4f85965f18b468"},"content":"tcpConnectOnce(host, port, timeoutMs) in src/preview/launch.ts (warren-49d9 / pl-592f step 1) is the phase-1 TCP-only probe. Uses Bun.connect with an open/error/timeout settle pattern; resolves 'connected' the instant the kernel completes the TCP handshake (immediately closes the socket via sock.end()) or 'not_connected' on refused/unreachable/timeout. Phase-2 readiness keeps using the HTTP probeOnce; phase-1 was misclassifying slow-first-headers dev servers (Next.js mid-compile) as not_connected and burning the connect budget. Exported (not file-local) so step-3 unit tests can drive it directly without spinning up attemptPreviewLaunch, and so noUnusedLocals stays happy during the multi-step plan rollout (step 2 wires it into the phase-1 loop).","id":"mx-ff4560"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-26T06:21:15.708Z","evidence":{"seeds":"warren-44ed","commit":"3f48d171612a4390081ed2305a3128e731cfa636"},"relates_to":["mx-ff4560"],"content":"launchPreview phase-1 (warren-44ed / pl-592f step 2) calls tcpConnectImpl = input.tcpConnect ?? tcpConnectOnce on '127.0.0.1':port — raw TCP handshake, not HTTP fetch. probeOnce stays in phase-2 only. Tests inject tcpConnect via the new optional LaunchPreviewInput.tcpConnect field; helpers alwaysConnected/alwaysRefused/fakeTcp in launch.test.ts cover the common test shapes. Real callers leave tcpConnect undefined so the real Bun.connect-based tcpConnectOnce runs.","id":"mx-db5bc8"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-26T06:26:26.491Z","evidence":{"commit":"19435f13140d4d4dcd1155f0b937b9f9d0d02d23"},"relates_to":["mx-ff4560"],"content":"tcpConnectOnce unit tests (src/preview/launch.test.ts, warren-f04c / pl-592f step 3) cover the real Bun.connect path on localhost: 'connected' against a Bun.listen server (asserts socket closes), 'not_connected' against a refused port (listener stopped first), and 'not_connected' against 192.0.2.1:1 (TEST-NET-1, dropped SYN) under timeoutMs. The slow-first-headers regression test in the two-phase describe block injects tcpConnect=[connected] + a fetch that throws DOMException AbortError on the first 2 calls then returns 200 — proves the warren-c3a2 bug stays fixed (phase-1 transitions immediately, phase-2 readiness budget absorbs the mid-headers aborts).","id":"mx-eb15b1"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-27T21:59:43.476Z","evidence":{"commit":"84d617c44f2cd303ec467d8cdb497947c5202de7"},"content":"src/preview/proxy.ts (warren-b902 / pl-9088 step 7) is split into src/preview/proxy/ subdirectory: route-match.ts (host/path/referer parsers + WARREN_API_PATH_PREFIXES), rewrite.ts (HTML <base> injection + root-relative href/src/srcset rewriting + Location: header rewriting + isHtmlContentType + HTML_HEAD_LOOKAHEAD_BYTES), forward.ts (forwardToUpstream + maybeFlushLastHit + DEFAULT_DEBOUNCE_MS), responses.ts (previewError + previewUnauthorized + LOGIN_PATH_PREFIX), types.ts (PreviewProxyConfig + Deps), index.ts (createPreviewProxyHandler orchestrator + public re-exports). Tests split alongside (route-match.test.ts, rewrite.test.ts, handler-subdomain.test.ts, handler-path.test.ts, handler-path-rewrites.test.ts, handler-path-referer.test.ts) with test-helpers.ts providing setupProxyEnv() to dedupe the in-memory db + repos + auth + live preview run boilerplate. External importers (src/server/main.ts, src/server/handlers/runs.preview.test.ts) now point at ../preview/proxy/index.ts. Removed proxy.ts + proxy.test.ts entries from scripts/file-size-budgets.json; replaced proxy.ts with proxy/index.ts + proxy/rewrite.ts in biome.json noExcessiveCognitiveComplexity override (matchAttributeAt and indexOfAsciiCaseInsensitive remain byte-level scanners that still trip the cognitive-complexity ceiling).","id":"mx-b3f2df"}
