+++ biome.json
@@ -100,7 +100,8 @@
 				"src/preview/cookie.ts",
 				"src/preview/eviction.ts",
 				"src/preview/launch.ts",
-				"src/preview/proxy.ts",
+				"src/preview/proxy/index.ts",
+				"src/preview/proxy/rewrite.ts",
 				"src/projects/refresh.ts",
 				"src/registry/refresh.test.ts",
 				"src/runs/burrow-config.ts",
