+++ scripts/file-size-budgets.json
@@ -24,8 +24,6 @@
 		"src/preview/eviction.ts": 819,
 		"src/preview/launch.test.ts": 995,
 		"src/preview/launch.ts": 712,
-		"src/preview/proxy.test.ts": 1593,
-		"src/preview/proxy.ts": 912,
 		"src/projects/manage.test.ts": 763,
 		"src/projects/refresh.test.ts": 709,
 		"src/registry/refresh.test.ts": 1041,
